Restructure application exports
