+++ arc_application/views/childminder_views/review.py
@@ -1,14 +1,10 @@
-import base64
 import time
 from datetime import datetime
 
-import requests
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.models import Group
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
-from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.views import View
 
@@ -21,16 +17,12 @@ from arc_application.magic_link import generate_magic_link
 from arc_application.notify import send_email
 from arc_application.views.base import release_application
 
-from django.core import serializers
-from ...messaging import SQSHandler
+from ...messaging import ApplicationExporter
 
 from ...models import ApplicantName, ApplicantPersonalDetails, Application, Arc, ArcComments, ChildcareType, ChildcareTraining, UserDetails, OtherPersonPreviousRegistrationDetails, PreviousName, \
-PreviousRegistrationDetails, PreviousAddress, HealthDeclarationBooklet, FirstAidTraining, Reference, CriminalRecordCheck, ChildInHome, ApplicantHomeAddress, AdultInHome, \
-HealthCheckHospital, HealthCheckSerious, HealthCheckCurrent
+PreviousRegistrationDetails, AdultInHome
 
 decorators = [login_required, group_required(settings.ARC_GROUP), user_assigned_application]
-
-sqs_handler = SQSHandler(settings.APPLICATION_QUEUE_NAME)
 
 
 @group_required(settings.ARC_GROUP)
@@ -58,6 +50,9 @@ def review(request):
             first_name = applicant_name.first_name
 
     if all_complete(application_id_local, True):
+
+        ApplicationExporter.export_childminder_application(application_id_local)
+
         personalisation = {'first_name': first_name, 'ref': app_ref}
         accepted_email(email, first_name, app_ref, application_id_local)
         send_survey_email(email, personalisation, application)
@@ -74,9 +69,6 @@ def review(request):
         variables = {
             'application_id': application_id_local,
         }
-
-        export = __create_full_application_export(application_id_local)
-        sqs_handler.send_message(export)
 
         return render(request, 'review-confirmation.html', variables)
 
@@ -110,16 +102,6 @@ def review(request):
         }
 
         return render(request, 'review-sent-back.html', variables)
-
-
-def has_group(user, group_name):
-    """
-    Check if user is in group
-    :return: True if user is in group, else false
-    """
-    group = Group.objects.get(name=group_name)
-
-    return True if group in user.groups.all() else False
 
 
 def send_survey_email(email, personalisation, application):
@@ -364,114 +346,3 @@ class OtherPersonPreviousRegistrationDetailsView(View):
             }
 
             return render(request, 'childminder_templates/add-previous-registration.html', context=variables)
-
-
-def __create_full_application_export(application_id):
-    """
-    Method for exporting a full application in a dictionary format
-    :param application_id: the identifier of the application to be exported
-    :return: a dictionary export of an application
-    """
-
-    export = {}
-
-    application = Application.objects.filter(application_id=application_id)
-    export['application'] = serializers.serialize('json', list(application))
-
-    adults_in_home = AdultInHome.objects.filter(application_id=application_id).order_by('adult')
-    adults_in_home_export = []
-    export['adults_in_home'] = serializers.serialize('json', list(adults_in_home))
-
-    # Iterate adults in home, appending prior names and addresses
-
-    for adult_in_home in adults_in_home:
-        previous_name = PreviousName.objects.filter(person_id=adult_in_home.pk)
-        previous_address = PreviousAddress.objects.filter(person_id=adult_in_home.pk)
-        serious_illness = HealthCheckSerious.objects.filter(person_id=adult_in_home.pk)
-        hospital_admissions = HealthCheckHospital.objects.filter(person_id=adult_in_home.pk)
-        current_illnesses = HealthCheckCurrent.objects.filter(person_id=adult_in_home.pk)
-
-        adults_in_home_export.append({
-            'adult': adult_in_home.adult,
-            'previous_names': serializers.serialize('json', list(previous_name)),
-            'previous_address': serializers.serialize('json', list(previous_address)),
-            'serious_illness': serializers.serialize('json', list(serious_illness)),
-            'hospital_admissions': serializers.serialize('json', list(hospital_admissions)),
-            'current_illnesses': serializers.serialize('json', list(current_illnesses))
-        })
-
-    export['health'] = adults_in_home_export
-
-    applicant_name = ApplicantName.objects.filter(application_id=application_id)
-    export['applicant_name'] = serializers.serialize('json', applicant_name)
-
-    applicant_personal_details = ApplicantPersonalDetails.objects.filter(application_id=application_id)
-    export['applicant_personal_details'] = serializers.serialize('json', list(applicant_personal_details))
-
-    applicant_home_address = ApplicantHomeAddress.objects.filter(application_id=application_id)
-    export['applicant_home_address'] = serializers.serialize('json', list(applicant_home_address))
-
-    child_in_home = ChildInHome.objects.filter(application_id=application_id)
-    export['child_in_home'] = serializers.serialize('json', list(child_in_home))
-
-    childcare_type = ChildcareType.objects.filter(application_id=application_id)
-    export['childcare_type'] = serializers.serialize('json', list(childcare_type))
-
-    criminal_record_check = CriminalRecordCheck.objects.filter(application_id=application_id)
-    export['criminal_record_check'] = serializers.serialize('json', list(criminal_record_check))
-
-    eyfs = ChildcareTraining.objects.filter(application_id=application_id)
-    export['eyfs'] = serializers.serialize('json', list(eyfs))
-
-    first_aid_training = FirstAidTraining.objects.filter(application_id=application_id)
-    export['first_aid_training'] = serializers.serialize('json', list(first_aid_training))
-
-    health_declaration_booklet = HealthDeclarationBooklet.objects.filter(application_id=application_id)
-    export['health_declaration_booklet'] = serializers.serialize('json', list(health_declaration_booklet))
-
-    previous_registration_details = PreviousRegistrationDetails.objects.filter(application_id=application_id)
-    export['previous_name'] = serializers.serialize('json', list(previous_registration_details))
-
-    references = Reference.objects.filter(application_id=application_id)
-    export['references'] = serializers.serialize('json', list(references))
-
-    user_details = UserDetails.objects.filter(application_id=application_id)
-    export['user_details'] = serializers.serialize('json', list(user_details))
-
-    # Create document exports
-
-    documents = {}
-
-    # Create endpoint URL location
-    eyc_endpoint = reverse('application_summary_pdf') + '?id=' + application_id
-
-    # Retrieve document
-    file_result = requests.get(eyc_endpoint, allow_redirects=True)
-    base64_string = str(base64.b64encode(file_result.content).decode("utf-8"))
-
-    documents['EYC'] = base64_string
-
-    if len(adults_in_home):
-
-        adult_documents = {}
-
-        for adult in adults_in_home:
-            endpoint = reverse('application_summary_pdf_adult') + '?id=' \
-                       + application_id + '&adult_id=' + adult.adult_id
-
-            # Retrieve document
-            file_result = requests.get(endpoint, allow_redirects=True)
-            base64_string = str(base64.b64encode(file_result.content).decode("utf-8"))
-
-            adult_document_object = {
-                "adult_id": adult.adult_id,
-                "document": base64_string,
-            }
-
-            adult_documents.update(adult_document_object)
-
-        documents['EY2'] = adult_documents
-
-    export['documents'] = documents
-
-    return export
